+++ src/configs/minimal.ts
@@ -9,11 +9,13 @@ export const minimal: Linter.Config = {
   },
   rules: {
     'tailwindcss/no-duplicate-imports': 'error',
+    'tailwindcss/no-duplicate-reference': 'error',
     'tailwindcss/no-empty-blocks': 'error',
     'tailwindcss/no-important': 'error',
     'tailwindcss/no-invalid-at-rules': 'error',
     'tailwindcss/no-invalid-named-grid-areas': 'error',
     'tailwindcss/no-invalid-properties': 'error',
+    'tailwindcss/require-reference-in-vue': 'error',
     'tailwindcss/use-baseline': 'error',
     'tailwindcss/valid-apply-directive': 'error',
     'tailwindcss/valid-modifier-syntax': 'error',