+++ src/configs/strict.ts
@@ -10,6 +10,7 @@ export const strict: Linter.Config = {
   rules: {
     'tailwindcss/consistent-spacing': 'error',
     'tailwindcss/no-duplicate-imports': 'error',
+    'tailwindcss/no-duplicate-reference': 'error',
     'tailwindcss/no-empty-blocks': 'error',
     'tailwindcss/no-important': 'error',
     'tailwindcss/no-invalid-at-rules': 'error',
@@ -25,6 +26,7 @@ export const strict: Linter.Config = {
       categories: ['colors', 'spacing', 'sizing', 'typography', 'borders', 'effects'],
     }],
     'tailwindcss/relative-font-units': 'error',
+    'tailwindcss/require-reference-in-vue': 'error',
     'tailwindcss/use-baseline': ['error', {
       strictness: 'newly',
     }],